change to only constants
i removed all data that does not have the string constant in its name so that i can narrow it down to just constants. also, i changed the name and formatting of the file for the const values, which in turn changed the one line of code where the input file is named.